+++ client/components/SingleGenre.js
