+++ client/components/main.js
@@ -16,6 +16,7 @@ import Sidebar from './sidebar';
  *  rendered out by the component's `children`.
  */
 const Main = (props) => {
+  console.log('props',props)
   const {children, handleClick, isLoggedIn} = props
 
   return (
