+++ client/reducers/ProductReducer.js
